fix(ci): satisfy script lint and coverage gates
Why
The first coverage repair exposed repository lint and max-lines failures in the newly covered selector tests and orchestration code, and the fresh Buildkite verify job rejected dynamic environment cleanup in the subprocess tests.

What
Use Bun environment cleanup, split selector orchestration into a dedicated module, keep the install-free selector closure explicit in the pipeline, and preserve the image-selector behavior while satisfying the repository line-count gate.

Verification
- bun --no-install run --cwd scripts lint (0 errors; 5 pre-existing warnings)
- bun --no-install run --cwd scripts test:coverage (273 pass, 0 fail; 92.02% functions, 91.13% lines)
- bun --no-install test ./.buildkite/scripts/select-main-pipeline.test.ts (27 pass, 0 fail)
- bun --no-install test ./.buildkite/scripts/select-image-targets.test.ts (48 pass, 0 fail)
- bunx --no-install turbo run typecheck --filter=@shepherdjerred/root-scripts
- bun --no-install .buildkite/scripts/validate-pipeline.ts
- bun --no-install .buildkite/scripts/validate-pipeline-release.ts
- bunx --no-install lefthook run pre-commit
