+++ .buildkite/scripts/select-image-targets.ts
@@ -61,8 +61,7 @@ const GLOBAL_IMAGE_INPUTS = [
   // The image lane writes the digest handoff every downstream release step
   // reads, so a change to the producer must rebuild every target once.
   ".buildkite/scripts/buildkite-handoff.ts",
-  // Visibility is part of the image publication contract; changing the
-  // reconciliation helper must rebuild every target before downstream pulls.
+  // Visibility reconciliation is part of the image publication contract.
   ".buildkite/scripts/ghcr-package-visibility.ts",
   ".buildkite/scripts/ghcr-public-access.ts",
   ".buildkite/scripts/image-targets.ts",
@@ -112,9 +111,8 @@ const TARGET_PATH_PREFIXES: Readonly<Record<string, readonly string[]>> = {
   ],
 };
 
-function pathMatchesPrefix(path: string, prefix: string): boolean {
-  return prefix.endsWith("/") ? path.startsWith(prefix) : path === prefix;
-}
+const pathMatchesPrefix = (path: string, prefix: string): boolean =>
+  prefix.endsWith("/") ? path.startsWith(prefix) : path === prefix;
 
 /** Optional changed-file contents the selector attributes precisely. */
 export type SelectorInputs = {
