+++ .buildkite/scripts/select-main-pipeline.ts
@@ -1,14 +1,43 @@
 #!/usr/bin/env bun
 
 import { fixedCorpusMode } from "./migration-core.ts";
+import {
+  runSelection as runSelectionForMain,
+  type SelectionDependencies,
+} from "./select-main-pipeline-selection.ts";
 
 type UnknownRecord = Record<string, unknown>;
-type PipelineStep = UnknownRecord;
-type PipelineDocument = {
+export type PipelineStep = UnknownRecord;
+export type PipelineDocument = {
   readonly agents: unknown;
   readonly env: unknown;
   readonly steps: readonly PipelineStep[];
 };
+
+export async function runSelection(
+  document: PipelineDocument,
+  dependencies?: SelectionDependencies,
+): Promise<number> {
+  return runSelectionForMain(
+    document,
+    dependencies ?? {
+      prepareBase,
+      writeChangedFiles: writeSelectorChangedFiles,
+      selectLanes,
+      recordSelectedSteps,
+      uploadPipeline,
+      annotateFallback,
+      deleteChangedFiles: async (path) => {
+        try {
+          await Bun.file(path).delete();
+        } catch (error: unknown) {
+          const reason = error instanceof Error ? error.message : String(error);
+          console.error(`WARN: could not delete ${path}: ${reason}`);
+        }
+      },
+    },
+  );
+}
 
 const STEP_LANE_REQUIREMENTS: Readonly<Record<string, readonly string[]>> = {
   verify: [],
@@ -190,22 +219,7 @@ export function pipelinePayload(
   return payload;
 }
 
-/**
- * `--replace` swaps the not-yet-started remainder of the build for the steps
- * uploaded, rather than appending them (`buildkite-agent` v3.134: "Replace the
- * rest of the existing pipeline with the steps uploaded. Jobs that are already
- * running are not removed").
- *
- * On the default branch `upload-pipeline.sh` uploads only `main-bootstrap.yml`,
- * so the build holds just the selector step when this runs and either mode
- * would schedule the same graph the first time. Replace is what makes a
- * REPEATED upload safe: if `ci-selector-base` is retried, a second run uploads
- * the same graph over the first instead of scheduling every step twice.
- * `ci-selector-base` is itself running at that point, so it is never removed
- * and the `depends_on` every rendered step carries stays resolvable — the
- * dependency shape is pinned by "renders stable selector dependencies and no
- * duplicate keys".
- */
+/** Upload the selected graph with `--replace`, making selector retries safe. */
 export function pipelineUploadArguments(
   changedFilesPath: string | undefined,
 ): string[] {
@@ -418,20 +432,7 @@ export async function uploadPipeline(
 
 export const SELECTED_STEPS_METADATA_KEY = "ci-selected-main-steps";
 
-/**
- * Publish the uploaded step keys so `annotate-build-summary.ts` knows which
- * steps exist in this build. Without it the annotator queries every step in
- * `summarySteps`, and `buildkite-agent step get` exits nonzero for a step the
- * selector omitted, failing the summary job on an otherwise green build. The
- * fallback path clears it back to empty: it uploads the complete graph, so
- * every step exists.
- *
- * Call this BEFORE the upload it describes. Buildkite appends uploaded steps,
- * so a failure after the graph is in the build cannot be retried and must not
- * fail the build; writing first means a failed write still falls back to the
- * complete graph. Every caller therefore records the set it is about to
- * upload, and the value never outlives the attempt that wrote it.
- */
+/** Record selected keys before upload so summary annotation has a truthful graph. */
 type MetadataWriter = (command: readonly string[]) => Promise<number>;
 
 export async function recordSelectedSteps(
@@ -483,101 +484,6 @@ export async function annotateFallback(
   ]);
   if (exitCode !== 0) {
     console.error("WARN: could not annotate main CI selector fallback");
-  }
-}
-
-type SelectionDependencies = {
-  readonly prepareBase: () => Promise<string>;
-  readonly writeChangedFiles: (base: string) => Promise<string>;
-  readonly selectLanes: (base: string) => Promise<Map<string, boolean>>;
-  readonly recordSelectedSteps: (
-    selected: ReadonlySet<string>,
-  ) => Promise<void>;
-  readonly uploadPipeline: (
-    document: PipelineDocument,
-    steps: readonly PipelineStep[],
-    changedFilesPath: string | undefined,
-  ) => Promise<void>;
-  readonly annotateFallback: (reason: string) => Promise<void>;
-  readonly deleteChangedFiles: (path: string) => Promise<void>;
-};
-
-const productionSelectionDependencies: SelectionDependencies = {
-  prepareBase,
-  writeChangedFiles: writeSelectorChangedFiles,
-  selectLanes,
-  recordSelectedSteps,
-  uploadPipeline,
-  annotateFallback,
-  deleteChangedFiles: async (path) => {
-    try {
-      await Bun.file(path).delete();
-    } catch (error: unknown) {
-      const reason = error instanceof Error ? error.message : String(error);
-      console.error(`WARN: could not delete ${path}: ${reason}`);
-    }
-  },
-};
-
-export async function runSelection(
-  document: PipelineDocument,
-  dependencies: SelectionDependencies = productionSelectionDependencies,
-): Promise<number> {
-  let steps: Map<string, PipelineStep> | undefined;
-  let changedFilesPath: string | undefined;
-  let uploaded = false;
-
-  try {
-    steps = mainSteps(document);
-    assertSelectionContract(steps);
-    const base = await dependencies.prepareBase();
-    changedFilesPath = await dependencies.writeChangedFiles(base);
-    const decisions = await dependencies.selectLanes(base);
-    const selected = selectedKeys(steps, decisions);
-    const rendered = renderSteps(steps, selected);
-    validateRenderedSteps(rendered);
-    // Record the selection BEFORE uploading. Buildkite appends uploaded steps,
-    // so once the graph is in the build no later step may fail: recording
-    // afterwards let a transient metadata write turn an already-runnable build
-    // into a hard failure. Recording first also keeps the metadata a truthful
-    // description of what gets uploaded, because a failure here still falls
-    // back to the complete graph.
-    await dependencies.recordSelectedSteps(selected);
-    await dependencies.uploadPipeline(document, rendered, changedFilesPath);
-    uploaded = true;
-    console.log(`Uploaded ${selected.size.toString()} selected main CI steps`);
-    return 0;
-  } catch (error) {
-    // Uploading a DIFFERENT graph after one already landed is unsafe even
-    // under `--replace`: replace drops only steps that have not started, so
-    // any selected step already running would survive and the complete graph
-    // would schedule its own copy alongside. Past that point the failure has
-    // to surface rather than fall back.
-    if (uploaded) throw error;
-    const reason = error instanceof Error ? error.message : String(error);
-    console.error(`WARN: ${reason}; falling back to the complete main graph`);
-    const rendered = renderFallbackSteps(document, steps, changedFilesPath);
-    validateRenderedSteps(rendered);
-    // Clear any selection the failed attempt recorded: an empty value is how
-    // the summary learns the complete graph was uploaded and every step
-    // exists. Best-effort, and deliberately not allowed to abort the fallback:
-    // this path exists to get a graph into the build when selection failed, so
-    // a metadata write must never be what leaves main with no steps at all. A
-    // stale value only mislabels rows in the summary annotation.
-    try {
-      await dependencies.recordSelectedSteps(new Set());
-    } catch (clearError) {
-      const detail =
-        clearError instanceof Error ? clearError.message : String(clearError);
-      console.error(`WARN: could not clear the recorded selection: ${detail}`);
-    }
-    await dependencies.uploadPipeline(document, rendered, changedFilesPath);
-    await dependencies.annotateFallback(reason);
-    return 0;
-  } finally {
-    if (changedFilesPath !== undefined) {
-      await dependencies.deleteChangedFiles(changedFilesPath);
-    }
   }
 }
